Separate waiter from Connection class file
