+++ utils/protocol/connection.py
@@ -2,28 +2,7 @@ import asyncio
 from utils.logger import logger
 from utils.protocol.data_converters import parse_protocol_message
 from utils.protocol.message import Message
-
-
-class Waiter:
-    def __init__(self, command: str):
-        self.command = command
-        self.event = asyncio.Event()
-        self._message = None
-
-    @property
-    def message(self) -> Message:
-        return self._message
-
-    @message.setter
-    def message(self, message: Message):
-        if self.event.is_set():
-            raise RuntimeError("Message already set")
-        self._message = message
-        self.event.set()
-
-    async def wait_message(self):
-        await self.event.wait()
-        return self.message
+from utils.protocol.waiter import Waiter
 
 
 class Connection:
